extract HoneycombInstrumentedComposable to our distro
include compose compile options
